+++ assets/app/admin/dashboard/index.js
@@ -23,6 +23,13 @@ angular.module('a2.admin.dashboard', [
 .controller('AdminDashboardCtrl', function($scope, $http, $q, $controller) {
 
     $scope.plots = $controller('AdminDashboardPlotterController', {'$scope':$scope});
+    
+    $scope.UserStatsExport = function(){
+        $http.get('/admin/all-users')
+        .success(function(data) {
+            download(data)
+        });
+    }
 
     $http.get('/admin/dashboard-stats')
         .success(function(data) {
@@ -70,5 +77,59 @@ angular.module('a2.admin.dashboard', [
         var value = $scope.settings[setting] == 'on' ? 'off' : 'on';
         return this.setSetting(setting, value);
     };
+    
+    function convertToCSV(objArray) {
+        var array = typeof objArray != 'object' ? JSON.parse(objArray) : objArray;
+        var str = '';
+    
+        for (var i = 0; i < array.length; i++) {
+            var line = '';
+            for (var index in array[i]) {
+                if (line != '') line += ','
+                line += array[i][index];
+            }
+            str += line + '\r\n';
+        }
+        return str;
+    }
+    
+    function exportCSVFile(headers, items, fileTitle) {
+        if (headers) {
+            items.unshift(headers);
+        }
+        // Convert Object to JSON
+        var jsonObject = JSON.stringify(items);
+        var csv = convertToCSV(jsonObject);
+        var exportedFilenmae = fileTitle + '.csv' || 'export.csv';
+        var blob = new Blob([csv], { type: 'text/csv;charset=utf-8;' });
+        if (navigator.msSaveBlob) { // IE 10+
+            navigator.msSaveBlob(blob, exportedFilenmae);
+        } else {
+            var link = document.createElement("a");
+            if (link.download !== undefined) {
+                var url = URL.createObjectURL(blob);
+                link.setAttribute("href", url);
+                link.setAttribute("download", exportedFilenmae);
+                link.style.visibility = 'hidden';
+                document.body.appendChild(link);
+                link.click();
+                document.body.removeChild(link);
+            }
+        }
+    }
+    
+    function download(data){
+      var headers = { name: 'Name',email: "Email" };
+      var itemsFormatted = [];
+      
+      // format the data
+      data.forEach((item) => {
+          itemsFormatted.push({
+            name: item.firstname + " " + item.lastname,
+            email: item.email
+          });
+      });
+      exportCSVFile(headers, itemsFormatted, 'users');
+    }
 })
 ;